requirements.txt
Django>=5.2
djangorestframework==3.14.0
whitenoise==6.6.0
django-filter==25.2
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework-stubs==3.16.7
django-request-id
ruff==0.14.14
redis==5.0.1
django-redis
python-decouple==3.8
drf-spectacular==0.27.1
gunicorn==21.2.0
python-json-logger
django-extensions
httpx>=0.27
